+++ GameEngine/GameBase/Status/StatusManager.h
@@ -2,18 +2,18 @@
 #include "GameEngine/Editor/GlobalVariables.h"
 
 struct Status {
-	float HP; // 体力
-	float STR;// 筋力
-	float VIT;// 生命力
-	float AGI;// 素早さ
+	int32_t HP; // 体力
+	int32_t STR;// 筋力
+	int32_t VIT;// 生命力
+	int32_t AGI;// 素早さ
 
 	void ReceiveDamage(Status attacker, float power) {
 		//  [(攻撃力/2) * 攻撃倍率] - [防御力/4] でダメージを計算する
-		float damage = ((attacker.STR / 2.0f) * power) - (VIT / 4.0f);
+		int32_t damage = int32_t((attacker.STR / 2.0f) * power) - int32_t(VIT / 4.0f);
 
 		// ダメージが負の値である場合、0に修正すること
 		// そのうちダメージ最低保証などがあればここらへんも修正する
-		if (damage < 0.0f) { damage = 0.0f; }
+		if (damage < 0.0f) { damage = 0; }
 
 		// ヒットポイントを減少させる
 		HP -= damage;
